requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==3.2
django-advanced-filters==2.0.0
django-cors-headers==4.1.0
django-filter==2.0.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-filters==0.2.2
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
flake8==5.0.4
idna==3.6
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
oauthlib==3.2.2
openapi-codec==1.3.2
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
simplejson==3.19.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.7
zipp==3.15.0
